Allow users to register an action
Why:

This change will allow users to say they did something. There is a list
that matches a regex to a task. If an event is sent that matches on of the words that represent an action (did, done, made, wrote), then the application will look for fragments like "blog post" and "friday talk". If a task is found then an event will be created for the user.

I also created a new class called `SlackClient`. This class wraps the slack framework's client, allowing us to define and call methods that make more sense in our domain.

This change addresses the need by:

* Creating a new handler for slack events.
* Creating the class `SlackClient`.
* Writing tests.
* Creating a small DSL to register tasks and their regex.