Change from Heroku to Elastic Beanstalk (EB)
Why:

Elastic Beanstalk (EB) is a service on top of AWS cloud. It is similar to Heroku in the way that allows to quickly deploy and manage applications. There are three reasons for this change:

1. I'm curious about how EB works and how I can use it.
2. Heroku free version does not work because we need to keep the rails application alive for slack's real-time API to work.
3. The free version is enough for us, and even the paid version is way cheaper than Heroku.

Apart from all the understanding the AWS ecosystem and configuring EB, the hardest part was to set up background jobs. There is nothing similar to Heroku's Scheduler. All the solutions I found online involved creating a complex system. I choose to go with cron jobs using the `whenever` gem (by the way, this would never have been an option on Heroku). Moving to EB also required changing to ruby 2.2.3. So no lonely operator.

I'm also fixing some issues on the tests.

This change addresses the need by:

* Adding configuration files for EB.
* Configuring `whenever`.
* Fixing the tests.